+++ e3nn_jax/__init__.py
@@ -52,12 +52,12 @@ from e3nn_jax._src.tensor_products import (
 from e3nn_jax._src.grad import grad
 from e3nn_jax._src.activation import scalar_activation, normalize_function
 from e3nn_jax._src.gate import gate
-from e3nn_jax._src.batchnorm import BatchNorm
-from e3nn_jax._src.dropout import Dropout
+from e3nn_jax._src.batchnorm import BatchNorm  # TODO remove
+from e3nn_jax._src.dropout import Dropout  # TODO remove
 from e3nn_jax._src.mlp_haiku import MultiLayerPerceptron  # TODO remove
 from e3nn_jax._src.graph_util import index_add, radius_graph, scatter_sum
 from e3nn_jax._src.reduced_tensor_product import reduced_tensor_product_basis, reduced_symmetric_tensor_product_basis
-from e3nn_jax._src.symmetric_tensor_product import SymmetricTensorProduct
+from e3nn_jax._src.symmetric_tensor_product import SymmetricTensorProduct  # TODO remove
 from e3nn_jax._src.s2grid import from_s2grid, to_s2grid, s2grid
 
 # make submodules flax and haiku available