+++ e3nn_jax/haiku.py
@@ -1,4 +1,7 @@
 from e3nn_jax._src.linear_haiku import Linear
 from e3nn_jax._src.mlp_haiku import MultiLayerPerceptron
+from e3nn_jax._src.batchnorm import BatchNorm
+from e3nn_jax._src.dropout import Dropout
+from e3nn_jax._src.symmetric_tensor_product import SymmetricTensorProduct
 
-__all__ = ["Linear", "MultiLayerPerceptron"]
+__all__ = ["Linear", "MultiLayerPerceptron", "BatchNorm", "Dropout", "SymmetricTensorProduct"]